refactor(openapi): split ContractTelemetryApi out of contracts.rs (#444, wave B)
Extract the 6 contract telemetry & history handlers (record_usage, get_usage,
submit_feedback, get_recipe_log, get_contract_events, get_feedback) — the
/contracts/:id/{usage,feedback,recipe-log,events} sub-resources — plus their
6 DTO tests into a new ContractTelemetryApi type in contract_telemetry.rs.

These handlers are the per-contract read/write data surface, fully decoupled
from ContractsApi: they depend only on Database methods and shared DTOs, with
zero references to contracts.rs-private helpers (check_spending_alert_and_notify
and create_stripe_checkout_session are used only by create_rental_request).

contracts.rs: 2244 -> 1747 lines (now under the 2k target). Registered in
tuple 1 of create_combined_api (13 -> 14; tuple 2 stays at 15).

Verified byte-identical: spec_snapshot guard PASS, hash unchanged
(de652956...), 187 paths / 327 schemas; all 6 moved endpoints present with
identical methods + ApiTags::Contracts. cargo build/clippy --tests -p api
clean (0 warnings); 28/28 contracts+contract_telemetry tests pass.
